# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zernio Gateway API",
    "version": "0.1.0",
    "x-gateway-contract-version": "1",
    "x-capability-contracts": [
      "social.publishing@1",
      "social.scheduling@1"
    ],
    "description": "Zeptly's gateway to Zernio. Implements Gateway Contract v1 (/v1/gateway, /v1/capabilities, /v1/connections) and the capability contracts Zernio backs in this gateway: Social Publishing v1 (/v1/social/publishing) and Social Scheduling v1. Provider identifiers never appear in responses. Authenticated with ZS1-HMAC-SHA256 service signatures (see docs/SECURITY.md)."
  },
  "components": {
    "securitySchemes": {
      "zeptlyServiceSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Zeptly-Signature",
        "description": "ZS1-HMAC-SHA256 signature. Also requires X-Zeptly-Caller, X-Zeptly-Timestamp and (workspace routes) X-Zeptly-Workspace-Id. See docs/SECURITY.md."
      }
    },
    "schemas": {
...
      "CreateConnectionRequestInput": {
        "type": "object",
        "properties": {
          "channel": {
            "description": "Channel to connect (for this gateway: a social network)",
            "type": "string",
            "enum": [
              "linkedin",
              "instagram",
              "facebook",
              "threads",
              "tiktok",
              "pinterest",
              "youtube",
              "bluesky"
            ]
          },
          "network": {
            "description": "Deprecated alias of `channel`",
            "type": "string",
            "enum": [
              "linkedin",
              "instagram",
              "facebook",
              "threads",
              "tiktok",
              "pinterest",
              "youtube",
              "bluesky"
            ]
          },
          "returnUrl": {
            "description": "Zeptly URL the end user's browser returns to (origin must be allow-listed)",
            "type": "string",
            "format": "uri"
          },
          "credentials": {
            "type": "object",
            "properties": {
              "handle": {
                "type": "string",
                "minLength": 3,
                "maxLength": 253
              },
              "appPassword": {
                "type": "string",
                "minLength": 8,
                "maxLength": 128
              }
            },
            "required": [
              "handle",
              "appPassword"
            ],
            "description": "Bluesky handle + app password (credentials strategy). Forwarded once to the provider, never stored or logged."
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Canonical error code (see GET /v1/gateway → errorCodes and docs/GATEWAY-CONTRACT.md)"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "GatewayChannelList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GatewayChannel"
            }
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
      "ConnectionReconcileSummary": {
        "type": "object",
        "properties": {
          "checked": {
            "type": "number"
          },
          "changed": {
            "type": "number"
          },
          "adopted": {
            "type": "number"
          }
        },
        "required": [
          "checked",
          "changed",
          "adopted"
        ],
        "additionalProperties": false
      },
...
      "SocialConnectionList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialConnection"
            }
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
      "CreateConnectionResponse": {
        "type": "object",
        "properties": {
          "provisioning": {
            "$ref": "#/components/schemas/ProvisioningSession"
          },
          "connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialConnection"
            }
          }
        },
        "required": [
          "provisioning",
          "connections"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/v1/connections/channels": {
      "get": {
        "summary": "Channels this gateway can provision, with connection strategies",
        "tags": [
          "connections"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayChannelList"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections": {
      "get": {
        "summary": "List the workspace's connections",
        "tags": [
          "connections"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "in": "query",
            "name": "channel",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "linkedin",
                "instagram",
                "facebook",
                "threads",
                "tiktok",
                "pinterest",
                "youtube",
                "bluesky"
              ]
            },
            "in": "query",
            "name": "network",
            "required": false,
            "description": "Deprecated alias of `channel`"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "connected",
                "degraded",
                "reauthorization_required",
                "disconnected"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialConnectionList"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Initiate provisioning of a provider account",
        "tags": [
          "connections"
        ],
        "description": "oauth_redirect / provider_managed: returns a provisioning session with `authorizationUrl`; send the end user's browser there. The provider redirects back through this gateway, which returns the browser to `returnUrl?provisioningId=…&status=…`. credentials (Bluesky): pass `credentials`; the connection is created synchronously.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequestInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateConnectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections/reconcile": {
      "post": {
        "summary": "Reconcile this workspace's connections with the provider (health, expiry, lost finalizations)",
        "tags": [
          "connections"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionReconcileSummary"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/connect/callback/{state}": {
      "get": {
        "summary": "Provider OAuth return (browser redirect; not called by Zeptly)",
        "tags": [
          "connections"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "in": "query",
            "name": "session",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "in": "query",
            "name": "error",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 20,
              "maxLength": 128
            },
            "in": "path",
            "name": "state",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}